+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## [1.0.0]
-Initial Release
+## [1.0.1] - 2023.03.16
 
-Added basic notification functionality like subscribe to address event, subscribe and unsubscribe to address event, list all subscriptions and webhooks.
+### Added
+
+- 13 new notification types to subscribe to
+
+## [1.0.0] - 2023.03.06
+
+### Added
+
+- Added basic notification functionality like subscribe to address event, subscribe and unsubscribe to address event, list all subscriptions and webhooks.